Update ProjectUpdateView tests
